Add static library workaround.
Cocoapods disallows transitive static libraries as a dependency with the
`use_frameworks!` flag.

This adds a new subspec that eliminates the transitive dependency
requirement.

For regular users, the installation is unchanged:
`pod 'Segment-Taplytics'` or (new) `pod 'Segment-Taplytics/Core'.

For users affected by the static library issue, the installation is:
```
pod 'Segment-Taplytics/StaticLibWorkaround'
pod 'Taplytics'
```
Note that since the transitive dependency is eliminated, users are
required to declare the Taplytics dependency directly as well.